bind: ignore private fields in structs
This patch adds a check so that the types of private fields aren't exported
to the wrapper.
This is useful if writing a 'wrapper' package that buffers the python code
from complex Go concepts that aren't yet wrapped for python (like pointers
and channels).
The code:
 _examples/wrapper/real_code.go and
 _examples/wrapper/pywrapper/wrapper_code.go
can be compiled with the command:
 gopy bind github.com/go-python/gopy/_examples/wrapper/pywrapper
and the functions manipulate Go classes gopy doesn't yet support, but they
are private, and not exported, so now the code will compile.

Fixes #60.